requirements.txt
flask
pika
#netifaces 
transformers
Pillow
requests
accelerate
tiktoken
einops
transformers_stream_generator==0.0.4
scipy
torchvision
pillow
tensorboard
matplotlib
bitsandbytes
optimum
auto-gptq
torch
